pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cs-binding-generator"
version = "0.1.0"
description = "Generate C# bindings from C headers using libclang with modern LibraryImport"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["csharp", "bindings", "interop", "libclang", "code-generation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: C",
    "Programming Language :: C#",
    "Topic :: Software Development :: Code Generators",
    "Topic :: Software Development :: Libraries",
]

dependencies = [
    "libclang>=16.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "mypy>=1.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/cs-binding-generator"
Repository = "https://github.com/yourusername/cs-binding-generator"
Issues = "https://github.com/yourusername/cs-binding-generator/issues"

[project.scripts]
cs-binding-generator = "cs_binding_generator.main:main"

[tool.setuptools]
packages = ["cs_binding_generator"]

[tool.setuptools.package-data]
cs_binding_generator = ["py.typed"]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true